move utils out of tweetPacs into helpers.js :wrench:
