+++ app/tweetPACs.js
@@ -1,12 +1,19 @@
 const async = require('async');
+const moment = require('moment');
+const config = require('../config');
 
 const propublicaService = require('../services/propublica');
-const config = require('../config');
 const twitterService = require('../services/twitter');
 const bitlyService = require('../services/bitly')
+
 const Vote = require('../models/vote');
 const PACContribution = require('../models/pacContribution');
-const moment = require('moment');
+
+const okayToTweet = require('../util/helpers').okayToTweet;
+const handleNullValues = require('../util/helpers').handleNullValues;
+const handleDonorName = require('../util/helpers').handleDonorName;
+const toProperCase = require('../util/helpers').toProperCase; 
+
 
 module.exports = tweetContribution = (cb) => {
     PACContribution.findOne({
@@ -35,7 +42,8 @@ module.exports = tweetContribution = (cb) => {
         twitterService.tweet(pacMessage, (err) => {
           if (err) {
             return cb(err);
-        }       
+        }
+        console.log('Tweeting PAC data:', pacMessage)       
         // save any new PAC data contribution entries to the database
         contribution.tweetedAt = new Date();
         contribution.save(cb);
@@ -47,65 +55,19 @@ module.exports = tweetContribution = (cb) => {
   })
 }
 
-// tweeting PAC contributions every 2 hours
-const okayToTweet = (contribution) => {
-  if (!contribution.tweetedAt) {
-    return true;
-  }
-  const tweetedAt = new moment(contribution.tweetedAt);
-  const cutoff = new moment().subtract(2, 'h');
-    if (tweetedAt.isBefore(cutoff)) {
-      return true;
-    }
-  return false;
-}
-
-// properCase helper
-String.prototype.toProperCase = function () {
-  return this.replace(/\w\S*/g, function(txt){return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();});
-}
-
-// handle null values for string methods
-const handleNullValues = (obj) => {
-  Object.keys(obj).forEach(function(key) {
-    if(obj[key] === null) {
-      obj[key] = '-';
-    }
-  })
-  return obj;
-}
-
-// to avoid 186 errors, we will have to abbreviate
-const handleDonorName = (str) => {
-  var mapObj = {
-   'Political Action Committee':"PAC",
-   'pac':"PAC",
-   'Pac':"PAC",
-   'political action committee':"PAC",
-   'Association':"Assn.",
-   'Companies':"Co's."
-  };
-
-  var re = new RegExp(Object.keys(mapObj).join("|"),"gi");
-  str = str.replace(re, function(matched){
-    return mapObj[matched];
-  });
-  return str
-}
-
 getPacTweetString = (contribution, cb) => {
   const name = config.congressPerson.name;
   const party = config.congressPerson.party;
   const jurisdiction = config.congressPerson.jurisdiction;
   const handle = config.congressPerson.handle;
-  const committee = contribution.committee.name.toProperCase();
-  const donor = contribution.donor_committee_name.toProperCase();
+  const committee = toProperCase(contribution.committee.name)
+  const donor = toProperCase(contribution.donor_committee_name)
   const abbrevDonor = handleDonorName(donor);
   const loadDate = moment(contribution.load_date).format('YYYY-MM-DD');
   const amount = contribution.contribution_receipt_amount.toLocaleString();
-  const donorDescription = contribution.entity_type_desc.toProperCase();
+  const donorDescription = toProperCase(contribution.entity_type_desc)
   const donorState = contribution.contributor_state;
-  const donorCity = contribution.contributor_city.toProperCase();
+  const donorCity = toProperCase(contribution.contributor_city)
   const pdf = contribution.pdf_url;
   
   bitlyService.shortenUrl(pdf, (err, shortUrl) => {
